- shunt operflags over to a long long - correct some bitmasks using integer when they should be using long
git-svn-id: http://svn.ratbox.org/svnroot/rserv/trunk@26859 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ include/conf.h
@@ -162,8 +162,8 @@ struct conf_oper
         char *host;
         char *pass;
 	char *server;
-	unsigned int flags;		/* general flags */
-	unsigned int sflags;		/* individual service flags */
+	unsigned long flags;		/* general flags */
+	unsigned long long sflags;	/* individual service flags */
 	int refcount;
 };
